feat: 장바구니 db 구현

+++ app/src/main/java/woowacourse/shopping/domain/repository/ProductRepository.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.domain.repository
 import woowacourse.shopping.domain.model.Product
 
 interface ProductRepository {
-    fun findProductById(id: Int): Result<Product>
+    fun findProductById(id: Long): Result<Product>
 
     fun getPagingProduct(
         page: Int,
