+++ app/src/main/java/woowacourse/shopping/presentation/ui/productlist/ProductListNavigateAction.kt
@@ -1,5 +1,5 @@
 package woowacourse.shopping.presentation.ui.productlist
 
 sealed interface ProductListNavigateAction {
-    data class NavigateToProductDetail(val productId: Int) : ProductListNavigateAction
+    data class NavigateToProductDetail(val productId: Long) : ProductListNavigateAction
 }
